Refactor DispatchChannel and Middleware Updates
Refactored `DispatchChannel` for improved performance and maintainability, replacing older queue implementations with modern constructs like `Channel`. Reorganized code into regions and added new methods for queue management and eviction policies.

Renamed `FrameDecompressMiddleware` to `DecompressMiddleware` and `FrameDecryptionMiddleware` to `DecryptMiddleware`. Updated `PacketDispatchOptions` to use the renamed middleware in the default pipeline.

Updated namespaces for consistency, cleaned up redundant code, and added XML documentation for clarity. Introduced logging capabilities in `PacketDispatchOptions` and adjusted access levels in `PacketMetadataProviders`.

Optimized critical methods with aggressive inlining and modern synchronization mechanisms for better runtime performance.
